+++ vm/zone.c
@@ -53,9 +53,9 @@ int _vm_zoneCreate(vm_zone_t *zone, size_t blocksz, unsigned int blocks)
 
 	/* Prepare zone for allocations */
 	for (i = 0; i < blocks; i++) {
-		*((void **)(zone->vaddr + i * blocksz)) = zone->vaddr + (i + 1U) * blocksz;
+		*((void **)(zone->vaddr + i * blocksz)) = zone->vaddr + ((size_t)i + 1U) * blocksz;
 	}
-	*((void **)(zone->vaddr + (blocks - 1U) * blocksz)) = NULL;
+	*((void **)(zone->vaddr + ((size_t)blocks - 1U) * blocksz)) = NULL;
 
 	zone->first = zone->vaddr;
 	zone->blocks = blocks;
